refactor build preset query function
