+++ src/shared/GmdVizPanel/types/timeseries/getTimeseriesQueryRunnerParams.ts
@@ -2,6 +2,7 @@ import { utf8Support } from '@grafana/prometheus';
 import { type SceneDataQuery } from '@grafana/scenes';
 import { promql } from 'tsqtsq';
 
+import { buildCustomFunctionQuery, buildPresetFunctionQuery } from 'shared/GmdVizPanel/buildFunctionQuery';
 import { buildQueryExpression } from 'shared/GmdVizPanel/buildQueryExpression';
 import { isRangeVectorFunction, PROMQL_FUNCTIONS, type PrometheusFunction } from 'shared/GmdVizPanel/config/promql-functions';
 import { QUERY_RESOLUTION } from 'shared/GmdVizPanel/config/query-resolutions';
@@ -83,51 +84,6 @@ function buildGroupByQueries({
   ];
 }
 
-// KG-supplied customFunction workaround (issue #1131). Range-vector functions need `[interval]`;
-// see isRangeVectorFunction for the detection rationale. Any function name is emitted verbatim.
-function buildCustomFunctionQuery(
-  metricName: string,
-  customFn: string,
-  expr: string,
-  interval: string,
-  isCounter: boolean
-): SceneDataQuery {
-  const isRange = isRangeVectorFunction(customFn);
-  const queryExpr = isRange ? `${customFn}(${expr}[${interval}])` : `${customFn}(${expr})`;
-  const fnName = isCounter ? `${customFn}(rate)` : customFn;
-  return {
-    refId: `${metricName}-${fnName}`,
-    expr: queryExpr,
-    legendFormat: fnName,
-    fromExploreMetrics: true,
-  };
-}
-
-// Preset-function path: looks up the fn in PROMQL_FUNCTIONS and applies it. Returns undefined
-// and logs a warn if the fn is not registered.
-function buildPresetFunctionQuery(
-  metricName: string,
-  fn: PrometheusFunction,
-  expr: string,
-  interval: string,
-  isCounter: boolean
-): SceneDataQuery | undefined {
-  const entry = PROMQL_FUNCTIONS.get(fn);
-  if (!entry) {
-    logger.warn(`[getTimeseriesQueryRunnerParams] Unknown PromQL function "${fn}", skipping query.`);
-    return undefined;
-  }
-  const isRangeFn = isRangeVectorFunction(entry.name);
-  const query = isRangeFn ? entry.fn({ expr, interval }) : entry.fn({ expr });
-  const fnName = isCounter ? `${entry.name}(rate)` : entry.name;
-  return {
-    refId: `${metricName}-${fnName}`,
-    expr: query,
-    legendFormat: fnName,
-    fromExploreMetrics: true,
-  };
-}
-
 // here we support preset functions
 function buildQueriesWithPresetFunctions({
   metric,
@@ -156,7 +112,7 @@ function buildQueriesWithPresetFunctions({
   const queryDefs: QueryDefs = queryConfig.queries?.length ? queryConfig.queries : [{ fn: defaultPromqlFn }];
   const queries: SceneDataQuery[] = [];
   for (const { fn } of queryDefs) {
-    const q = buildPresetFunctionQuery(metric.name, fn, expr, interval, isCounter);
+    const q = buildPresetFunctionQuery(metric.name, fn, expr, interval, isCounter, '[getTimeseriesQueryRunnerParams]');
     if (q) {
       queries.push(q);
     }
